completed team activity 3
