+++ src/js/checkout.js
@@ -2,4 +2,3 @@ import { loadHeaderFooter } from "./utils.mjs";
 
 //Loade Header and Footer
 loadHeaderFooter();
-
