+++ src/public/partials/footer.html
@@ -1,4 +1,4 @@
 <p>
-    &copy;2025 ⛺ SleepOutside ⛺ WDD 330 ⛺ BYU-Idaho for BYU-Pathway
-    Worldwide Online
+  &copy;2025 ⛺ SleepOutside ⛺ WDD 330 ⛺ BYU-Idaho for BYU-Pathway Worldwide
+  Online
 </p>
